chore: Merge Decomposition and Anonymization into one command
