--- Modules/Relationships/src/Relationships.Application/Relationships/Commands/AnonymizeRelationshipsOfIdentity/AnonymizeRelationshipsOfIdentityCommand.cs
@@ -1,13 +1,0 @@
-﻿using MediatR;
-
-namespace Backbone.Modules.Relationships.Application.Relationships.Commands.AnonymizeRelationshipsOfIdentity;
-
-public class AnonymizeRelationshipsOfIdentityCommand : IRequest
-{
-    public AnonymizeRelationshipsOfIdentityCommand(string identityAddress)
-    {
-        IdentityAddress = identityAddress;
-    }
-
-    public string IdentityAddress { get; set; }
-}
