+++ Modules/Relationships/src/Relationships.Application/Relationships/Commands/DecomposeAndAnonymizeRelationshipsOfIdentity/DecomposeAndAnonymizeRelationshipsOfIdentityCommand.cs
@@ -1,0 +1,13 @@
+using MediatR;
+
+namespace Backbone.Modules.Relationships.Application.Relationships.Commands.DecomposeAndAnonymizeRelationshipsOfIdentity;
+
+public class DecomposeAndAnonymizeRelationshipsOfIdentityCommand : IRequest
+{
+    public DecomposeAndAnonymizeRelationshipsOfIdentityCommand(string identityAddress)
+    {
+        IdentityAddress = identityAddress;
+    }
+
+    public string IdentityAddress { get; set; }
+}
